[Refactor] Add test to each reducer
